* Cambio de nombre y ubicación de las clases de bienes, categorias y sus atributos.
* Cambio de nombre para atributo foto->fotoUrl de la clase bien
