--- donaciones-service/src/main/java/Bienes/Atributos/Estado.java
@@ -1,8 +1,0 @@
-package Bienes.Atributos;
-
-public enum Estado {
-    NUEVO,
-    USADO,
-    DESGASTADO
-}
-
