--- donaciones-service/src/main/java/Bienes/Atributos/Subcategoria.java
@@ -1,14 +1,0 @@
-package Bienes.Atributos;
-
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-
-public class Subcategoria {
-    private Categoria categoria;
-    private String nombre;
-
-}
