--- donaciones-service/src/main/java/Bienes/Atributos/Unidad.java
@@ -1,10 +1,0 @@
-package Bienes.Atributos;
-
-public enum Unidad {
-    UNIDADES,
-    KILOGRAMO,
-    LITROS,
-    METRO,
-    METRO_CUADRADO,
-    METRO_CUBICO
-}
