+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/Bienes/Atributos/estado.java
@@ -1,0 +1,8 @@
+package ar.edu.utn.frba.ddsi.donaciones.models.entities.Bienes.Atributos;
+
+public enum estado {
+    NUEVO,
+    USADO,
+    DESGASTADO
+}
+
