+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/Bienes/Atributos/subCategoria.java
@@ -1,0 +1,14 @@
+package ar.edu.utn.frba.ddsi.donaciones.models.entities.Bienes.Atributos;
+
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+
+public class subCategoria {
+    private categoria categoria;
+    private String nombre;
+
+}
