+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/Bienes/Atributos/unidad.java
@@ -1,0 +1,10 @@
+package ar.edu.utn.frba.ddsi.donaciones.models.entities.Bienes.Atributos;
+
+public enum unidad {
+    UNIDADES,
+    KILOGRAMO,
+    LITROS,
+    METRO,
+    METRO_CUADRADO,
+    METRO_CUBICO
+}
